policy v5: continuous, confidence-driven decimal unit sizing
Units are no longer 1/2/3 buckets. size_units is now continuous in [1.0, 3.0]
(one decimal): conviction = prob_score · conf_score (how far into the strong-
favorite range × data depth), mapped through a gamma>1 curve so multi-unit is
sparing. Curve: 0.82→1.0u, 0.88→1.4u, 0.90→1.7u, 0.93→2.2u, 0.97+→3.0u; thin
data (low model_confidence) pins any pick to 1.0u; a suspicious edge stays 1.0u.
paper_bets.units migrated Integer→Float (migration b2f4a7c9d310, applied). All
unit displays formatted to 1 decimal. POLICY_VERSION v5.

+++ bot/models.py
@@ -288,7 +288,7 @@ class PaperBet(Base):
     model_confidence: Mapped[float] = mapped_column(Float)
     edge: Mapped[float] = mapped_column(Float)
     basis: Mapped[str] = mapped_column(String(16))  # 'prematch' | 'advisory'
-    units: Mapped[int] = mapped_column(Integer, default=1)  # 1-3; 3 is rare
+    units: Mapped[float] = mapped_column(Float, default=1.0)  # 1.0-3.0, decimal; multi-unit rare
     tier: Mapped[str | None] = mapped_column(String(8))
     state_at_placement: Mapped[str | None] = mapped_column(String(16))
     reasoning: Mapped[dict | None] = mapped_column(JSONB)  # snapshot for tuning
